// src/constants/roles.enum.ts
export enum Role {
  ADMIN = 'ADMIN', // ሊቀ መንበር (Admin)
  VICE_CHAIR = 'VICE_CHAIR', // ምክትል ሊቀ መንበር (Vice-Chair)
  SECRETARY = 'SECRETARY', // ዋና ጸሐፊ (Secretary)
  TREASURER = 'TREASURER', // ገንዘብ ያዥ (Treasurer)
  ACCOUNTANT = 'ACCOUNTANT', // ሒሳብ ሹም (Accountant)
  AUDITOR = 'AUDITOR', // ሒሳብ አጣሪ (Auditor)
  PROPERTY_ADMIN = 'PROPERTY_ADMIN', // ንብረት ጠባቂ (Property Admin)
  LAITY = 'LAITY', // ወጣት (Laity)
  PRIEST = 'PRIEST', // ካሊስ (Priest)
}